Create lab_7_1.c
Lab 7
Q1 - Write a program that takes 10 floating point numbers from the user as input and stores them in an float array, and computes the maximum element, second maximum element, minimum element, and second minimum element of an array. Then it computes the average of the four elements computed above and displays it to the user.