Create lab_11_3.c
Lab Sheet 11 - Q3
Write a program that takes two arrays of the same size as input. The size of the arrays is
not known till runtime and has to be taken from the user. The program then takes the
elements of the arrays from the user as input too. Finally, the program prints the dot product
of the two arrays on the console. [Taking dot product of the arrays is same as multiplying the
arrays element wise and then taking the sum of all the products]
Note that you need to declare all the arrays dynamically and allocate them on the heap for the
implementation of the above question, no static arrays are allowed.